Moving internal code to InternalNano and rename MapUtil to MapFactories.
